style: 반응형 - 식물 백과 목록 페이지
