+++ src/app/(greeny)/(home)/Home.module.scss
@@ -3,7 +3,7 @@
 .root {
   @include flex-vertical(0);
   position: relative;
-  max-width: 36rem;
+  // max-width: 36rem;
   width: 100%;
   overflow: hidden;
   min-height: 100vh;
