+++ src/app/(greeny)/books/Filter.module.scss
@@ -1,0 +1,186 @@
+@import '@styles/mixin';
+
+.filter_wrapper {
+  @include flex-vertical(3rem);
+  justify-content: space-between;
+  background: var(--color-white);
+  overflow-x: hidden;
+  margin: 0 auto;
+  max-width: 31.2rem;
+  .form {
+    @include flex-vertical(2rem);
+    .filter_group {
+      @include flex-vertical(3rem);
+      height: 40rem;
+      overflow-y: scroll;
+      &::-webkit-scrollbar {
+        width: 0.4rem;
+      }
+
+      &::-webkit-scrollbar-thumb {
+        background-color: var(--color-primary);
+        border-radius: 1rem;
+      }
+
+      .plant_type {
+        @include flex-vertical(2rem);
+        justify-content: space-between;
+        align-items: flex-start;
+        height: 8.5rem;
+        p {
+          font-weight: var(--font-semibold);
+          font-size: 1.4rem;
+        }
+        .label_wrapper {
+          display: grid;
+          grid-template-columns: repeat(3, 1fr);
+          gap: 1rem;
+          place-items: start start;
+        }
+      }
+
+      .plant_props {
+        @include flex-vertical(2rem);
+        justify-content: space-between;
+        align-items: flex-start;
+        p {
+          font-weight: var(--font-semibold);
+          font-size: 1.4rem;
+        }
+        .flower_color {
+          @include flex-vertical(1rem);
+          justify-content: space-between;
+          align-items: flex-start;
+          p {
+            font-weight: var(--font-semibold);
+            font-size: 1.4rem;
+          }
+          .label_wrapper {
+            display: grid;
+            grid-template-columns: repeat(3, 1fr);
+            gap: 1rem;
+            place-items: start start;
+            font-weight: var(--font-reqular);
+            font-size: 1.2rem;
+          }
+        }
+        .fruit_color {
+          @include flex-vertical(1rem);
+          justify-content: space-between;
+          align-items: flex-start;
+          p {
+            font-weight: var(--font-semibold);
+            font-size: 1.4rem;
+          }
+          .label_wrapper {
+            display: grid;
+            grid-template-columns: repeat(3, 1fr);
+            gap: 1rem;
+            place-items: start start;
+            font-weight: var(--font-reqular);
+            font-size: 1.2rem;
+          }
+        }
+        .leaf_pattern {
+          @include flex-vertical(1rem);
+          justify-content: space-between;
+          align-items: flex-start;
+          width: 100%;
+          p {
+            font-weight: var(--font-semibold);
+            font-size: 1.4rem;
+          }
+          .label_wrapper {
+            width: 25rem;
+            display: grid;
+            grid-template-columns: 1fr auto;
+            align-items: center;
+            row-gap: 1rem;
+            column-gap: 1rem;
+            font-weight: var(--font-reqular);
+            font-size: 1.2rem;
+          }
+        }
+      }
+
+      .light_quantity {
+        @include flex-vertical(2rem);
+        justify-content: space-between;
+        align-items: flex-start;
+        p {
+          font-weight: var(--font-semibold);
+          font-size: 1.4rem;
+        }
+        .label_wrapper {
+          width: 25rem;
+          display: grid;
+          grid-template-columns: 1fr 1fr;
+          row-gap: 1rem;
+          column-gap: 1rem;
+        }
+      }
+      .waterCycle {
+        @include flex-vertical(2rem);
+        justify-content: space-between;
+        p {
+          font-weight: var(--font-semibold);
+          font-size: 1.4rem;
+        }
+        .label_wrapper {
+          @include flex-vertical(1rem);
+          justify-content: space-between;
+          align-items: flex-start;
+        }
+      }
+    }
+
+    .btn_group {
+      @include flex-horizontal(1rem);
+      justify-content: center;
+      .btn_submit {
+        @include flex-horizontal(0);
+        justify-content: center;
+        align-items: center;
+        width: 7rem;
+        height: 3rem;
+        background: var(--color-primary);
+        border-radius: 1rem;
+        color: var(--color-white);
+      }
+      .btn_cancel {
+        @include flex-horizontal(0);
+        justify-content: center;
+        align-items: center;
+        width: 7rem;
+        height: 3rem;
+        background: var(--color-white);
+        border-radius: 1rem;
+        border: 0.1rem solid var(--color-primary);
+        box-sizing: border-box;
+      }
+    }
+  }
+}
+.custom_checkbox {
+  @include flex-horizontal(1rem);
+  & > input {
+    margin: 0;
+    appearance: none;
+    cursor: pointer;
+    width: 1.6rem;
+    height: 1.6rem;
+    border-radius: 0.35rem;
+    border: 0.1rem solid var(--color-primary);
+    background-image: url('/images/CheckBox_nor.svg');
+    background-size: 1.4rem 1.4rem;
+    background-position: center;
+    background-repeat: no-repeat;
+
+    &:checked {
+      background-image: url('/images/CheckBox_sel.svg');
+    }
+    & + span {
+      cursor: pointer;
+    }
+  }
+}
