+++ src/app/(greeny)/books/Modal.module.scss
@@ -16,7 +16,6 @@
     background: white;
     padding: 3rem 2rem 2rem;
     border-radius: 8px;
-    max-width: 31.2rem;
     width: 100%;
     position: relative;
 
@@ -29,5 +28,17 @@
       font-size: 1.5rem;
       cursor: pointer;
     }
+  }
+}
+
+@media (max-width: 48rem) {
+  .modal_content {
+    max-width: 31.2rem;
+  }
+}
+
+@media (min-width: 48rem) {
+  .modal_content {
+    max-width: 66.4rem;
   }
 }
